Move some common options to datasource; poster CTA fixup for right side
