+++ src/components/Container/Container.styles.ts
@@ -8,48 +8,3 @@ export const bgColors = {
   black: 'su-bg-black su-text-black-10',
   white: 'su-bg-white su-text-black',
 };
-
-export const paddingTops = {
-  none: '',
-  base: 'su-rs-pt-0',
-  1: 'su-rs-pt-1',
-  2: 'su-rs-pt-2',
-  3: 'su-rs-pt-3',
-  4: 'su-rs-pt-4',
-  5: 'su-rs-pt-5',
-  6: 'su-rs-pt-6',
-  7: 'su-rs-pt-7',
-  8: 'su-rs-pt-8',
-  9: 'su-rs-pt-9',
-  10: 'su-rs-pt-10',
-};
-
-export const paddingBottoms = {
-  none: '',
-  base: 'su-rs-pb-0',
-  1: 'su-rs-pb-1',
-  2: 'su-rs-pb-2',
-  3: 'su-rs-pb-3',
-  4: 'su-rs-pb-4',
-  5: 'su-rs-pb-5',
-  6: 'su-rs-pb-6',
-  7: 'su-rs-pb-7',
-  8: 'su-rs-pb-8',
-  9: 'su-rs-pb-9',
-  10: 'su-rs-pb-10',
-};
-
-export const paddingVerticals = {
-  none: '',
-  base: 'su-rs-py-0',
-  1: 'su-rs-py-1',
-  2: 'su-rs-py-2',
-  3: 'su-rs-py-3',
-  4: 'su-rs-py-4',
-  5: 'su-rs-py-5',
-  6: 'su-rs-py-6',
-  7: 'su-rs-py-7',
-  8: 'su-rs-py-8',
-  9: 'su-rs-py-9',
-  10: 'su-rs-py-10',
-};
