+++ src/components/Container/Container.tsx
@@ -1,14 +1,20 @@
 import React, { HTMLAttributes } from 'react';
 import { dcnb } from 'cnbuilder';
+import {
+  paddingTops,
+  paddingBottoms,
+  paddingVerticals,
+  PaddingType,
+} from '../../utilities/datasource';
 import * as styles from './Container.styles';
 import * as types from './Container.types';
 
 export type ContainerProps = HTMLAttributes<HTMLElement> & {
   as?: types.ContainerElementType;
   width?: types.WidthType;
-  pt?: types.PaddingType;
-  pb?: types.PaddingType;
-  py?: types.PaddingType;
+  pt?: PaddingType;
+  pb?: PaddingType;
+  py?: PaddingType;
   bgColor?: types.BgColorType;
   style?: React.CSSProperties;
 };
@@ -30,9 +36,9 @@ export const Container = ({
     style={style}
     className={dcnb(
       bgColor ? styles.bgColors[bgColor] : '',
-      py ? styles.paddingVerticals[py] : '',
-      pt ? styles.paddingTops[pt] : '',
-      pb ? styles.paddingBottoms[pb] : '',
+      py ? paddingVerticals[py] : '',
+      pt ? paddingTops[pt] : '',
+      pb ? paddingBottoms[pb] : '',
       width ? styles.widths[width] : '',
       className,
     )}
